requirements.txt
aiogram
requests
python-dotenv
openpyxl
sqlalchemy
psycopg2-binary
